requirements.txt
numpy
opencv-python
PyYAML
sounddevice
picamera2; platform_system == "Linux"
